+++ src/main/java/com/jobnote/domain/user/listener/VerificationEmailEventListener.java
@@ -3,8 +3,8 @@ package com.jobnote.domain.user.listener;
 import com.jobnote.domain.user.domain.VerificationToken;
 import com.jobnote.domain.user.dto.SignUpEvent;
 import com.jobnote.global.config.properties.AppProperties;
-import com.jobnote.mail.MailService;
-import com.jobnote.mail.dto.MailMessageDto;
+import com.jobnote.infra.mail.dto.MailMessageDto;
+import com.jobnote.infra.mail.service.MailService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.event.TransactionPhase;
